+++ CHANGELOG.md
@@ -19,6 +19,7 @@
     - add new `scrollbar` widget that connects to a scrolling container widget and acts as it scrollbar
     - input: add `noCanvas` option to avoid blurry text under some circumstances
     - xy: add `clipX` and `clipY` proprties (allows multixy to set per-pads anges limits)
+    - button: add `onTouch` script
 
 - editor
     - selecting a widget automatically shows it in the widget tree if it's not minimized